requirements.txt
streamlit>=1.36.0
plotly>=5.22.0
pandas>=2.2.0
statsmodels>=0.14.2  # for plotly express trendline option
openai>=1.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
